Remove Order destination() getter
Why not just use a destination() getter, that will get the destination country from the Customer? A problem arises if, after Order's submission, the Customer updates their "selectedAddress" field. Upon getting the Order from persistece (and, together with it, the Order's Customer), the destination will be updated accordingly, which SHOULDN'T happen. So we have to save the relevant bits of Customer's state at the time of Order submission.